add types to q_a and rename files to camel case
