+++ src/components/linkTeaser/dataService.ts
