+++ src/components/richTextCharLimit1000/dataService.ts
@@ -1,7 +1,7 @@
 import { ParagraphResourceType } from '@/types/paragraph'
 import { EntityMetaInfo } from '@/lib/delegators/entityMetaProvider'
 import { wysiwygDataService } from '@/components/wysiwyg/dataService'
-import { RichTextCharLimit1000 } from './'
+import { RichTextCharLimit1000 } from '.'
 
 export const Meta: EntityMetaInfo = {
   resource: ParagraphResourceType.RichTextCharLimit1000,
