+++ src/components/richTextCharLimit1000/index.tsx
