+++ src/pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from 'next/app'
 import type { ReactElement, ReactNode } from 'react'
 import { useEffect } from 'react'
 import { defineCustomElements } from '@department-of-veterans-affairs/web-components/loader'
-import { TAG_MANAGER_ARGS } from '@/lib/google-analytics'
+import { TAG_MANAGER_ARGS } from '@/lib/analytics'
 import TagManager from 'react-gtm-module'
 import route from 'next/dist/server/router'
 import '@/styles/globals.css'
