+++ src/pages/demo/linkTeasers/index.tsx
@@ -3,7 +3,7 @@ import { GetStaticPropsContext, GetStaticPropsResult } from 'next'
 import { DrupalJsonApiParams } from 'drupal-jsonapi-params'
 import Container from '@/components/container'
 import { ParagraphLinkTeaser, ParagraphResourceType } from '@/types/paragraph'
-import { LinkTeaser } from '@/components/link_teaser'
+import { LinkTeaser } from '@/components/linkTeaser'
 import { generalEntityDataService } from '@/lib/delegators/generalEntityDataService'
 const linkTeaserParams = { boldTitle: false, sectionHeader: '' }
